Change parent mode to special-mode.
This seems like the right choice, referring to the Elisp manual, which says that
special-mode is the mode to inherit from if the buffer displays text generated
programmatically by Emacs and not representing a file.

Since special-mode automatically makes the buffer read-only, we no longer need
to call `read-only-mode` explicitly.

Caveat: I haven't fully investigated the implications of special-mode binding
the `q` and `g` keys, but I don't think it should make any visible difference.